Broke unit-tests into individual tests for each subcomponent. Created a mock-data folder for mock-data and mock-hooks files used by multiple tests
