+++ jest.config.js
@@ -1,3 +1,4 @@
 module.exports = {
   collectCoverage: true,
+  testEnvironment: 'jsdom',
 };
